Split up block face, direction, and rotation
