+++ cosmos_client/src/structure/planet/align_player.rs
@@ -4,7 +4,7 @@ use std::f32::consts::PI;
 
 use bevy::prelude::{App, Commands, Component, Entity, Parent, Quat, Query, Transform, Update, Vec3, With, Without};
 use cosmos_core::{
-    block::BlockFace,
+    block::block_face::BlockFace,
     netty::client::LocalPlayer,
     physics::{gravity_system::GravityEmitter, location::Location},
     structure::{planet::Planet, ship::pilot::Pilot},
