+++ cosmos_core/src/structure/dynamic_structure.rs
@@ -12,7 +12,7 @@ use bevy::{
 use serde::{Deserialize, Serialize};
 
 use crate::{
-    block::{blocks::AIR_BLOCK_ID, Block, BlockRotation},
+    block::{block_rotation::BlockRotation, blocks::AIR_BLOCK_ID, Block},
     ecs::NeedsDespawned,
     events::block_events::BlockChangedEvent,
     registry::{identifiable::Identifiable, Registry},
