+++ cosmos_server/src/structure/station/loading.rs
@@ -2,7 +2,7 @@
 
 use bevy::prelude::{in_state, App, Commands, Component, Entity, EventWriter, IntoSystemConfigs, Query, Res, Update, With};
 use cosmos_core::{
-    block::{Block, BlockRotation},
+    block::{block_rotation::BlockRotation, Block},
     registry::Registry,
     structure::{
         coordinates::BlockCoordinate,
